+++ R/compute_outliers.R
@@ -14,7 +14,7 @@
 #' @inheritParams predict_bmi
 #' @param from A string indicating the type of data to be used for the AP and AR
 #'   computation, either "predicted" or "observed". Default is "predicted".
-#' @param outlier_method The outlier detection method(s). Default is `"iqr"`. Can be `"all"` or some of
+#' @param outlier_method The outlier detection method(s). Default is `"iqr"`. Can be
 #'   `"cook"`, `"pareto"`, `"zscore"`, `"zscore_robust"`, `"iqr"`, `"ci"`, `"eti"`,
 #'   `"hdi"`, `"bci"`, `"mahalanobis"`, `"mahalanobis_robust"`, `"mcd"`, `"ics"`,
 #'   `"optics"` or `"lof"`.
@@ -68,6 +68,14 @@ compute_outliers <- function(
 ) {
   value <- what <- AP <- AR <- NULL # no visible binding for global variable from data.table
   from <- match.arg(from, c("predicted", "observed"))
+  outlier_method <- match.arg(
+    arg = outlier_method,
+    choices = c(
+      "cook", "pareto", "zscore", "zscore_robust", "iqr", "ci",
+      "eti", "hdi", "bci", "mahalanobis", "mahalanobis_robust",
+      "mcd", "ics", "optics", "lof"
+    )
+  )
   apar_dt <- compute_apar(
     fit = fit,
     from = from,